Modification : Systeme de messagerie modifier, le message est envoyé, avec L'id du rental et du user connecté

+++ Back-End/SpringSecurityAuth/src/main/java/com/chatop/api/domain/entity/Message.java
@@ -15,10 +15,10 @@ public class Message {
     private Long id;
 
     @Column(name = "rental_id")
-    private int rentalId;
+    private Long rentalId;
 
     @Column(name = "user_id")
-    private int userId;
+    private Long userId;
 
     private String message;
